+++ api/index.py
@@ -396,6 +396,7 @@ _STATIC_EXACT = frozenset({
     "ui-v110.css",
     "ui-v213.css",
     "ui-v244.css",
+    "ui-v245.css",
     "sw.js",
     "manifest.webmanifest",
     # Login shows the running version + DE/EL "what changed" from this.